size_t for ndof, NodeId

+++ comp/hcurlhdivfes.hpp
@@ -50,9 +50,9 @@ public:
   virtual void UpdateCouplingDofArray();
 
   ///
-  virtual int GetNDof () const throw();
+  virtual size_t GetNDof () const throw();
   ///
-  virtual int GetNDofLevel (int level) const;
+  virtual size_t GetNDofLevel (int level) const;
 
   /// 
   virtual void GetDofRanges (ElementId ei, Array<IntRange> & dranges) const;
@@ -207,9 +207,9 @@ public:
   virtual void Update(LocalHeap & lh);
 
   ///
-  virtual int GetNDof () const throw();
+  virtual size_t GetNDof () const throw();
   ///
-  virtual int GetNDofLevel (int level) const;
+  virtual size_t GetNDofLevel (int level) const;
 
   ///
   virtual void GetDofNrs (ElementId ei, Array<int> & dnums) const;
